+++ macos/Classes/bundle_data.c
@@ -1,0 +1,1 @@
+#include "../../native/shim/bundle_data.c"