+++ models/Convert.java
@@ -1,7 +1,0 @@
-package io.quicktype;
-
-import com.fasterxml.jackson.annotation.*;
-
-@JsonAutoDetect(fieldVisibility=JsonAutoDetect.Visibility.NONE)
-public class Convert {
-}