+++ models/test.scala
@@ -1,8 +1,0 @@
-package quicktype
-
-case class Test (
-    val asdf : String,
-    val asdf2 : String
-)
-
-case class Convert()